pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pydeck>=0.9.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]
